+++ apps/frontend/src/libs/constants/numbers.ts
@@ -1,5 +1,3 @@
-const ZERO = 0;
-const ONE = 1;
 const LAST_INDEX = -1;
 
-export { LAST_INDEX, ONE, ZERO };
+export { LAST_INDEX };