+++ packages/shared/src/modules/quiz/libs/validation-schemas/quiz-submit-answers.validation-schema.ts
@@ -1,26 +1,16 @@
-import { sanitizeTextInput } from "shared";
 import z from "zod";
 
-import { QuizCategory, QuizValidationMessage } from "../enums/enums.js";
+import {
+	QuizAnswersRule,
+	QuizCategory,
+	QuizValidationMessage,
+} from "../enums/enums.js";
+import { quizAnswerSchema } from "./quiz-shared.validation-schema.js";
 
-const ZERO_ARRAY_LENGTH = 0;
-
-const QuizAnswersSchema = z.object({
-	answers: z.array(
-		z
-			.object({
-				isSkipped: z.boolean(),
-				questionId: z.number().int().positive(),
-				selectedOptions: z.array(
-					z.union([z.number().int().positive(), z.string().trim()]),
-				),
-				userInput: z.preprocess(
-					(value) =>
-						typeof value === "string" ? sanitizeTextInput(value) : "",
-					z.string().default(""),
-				),
-			})
-			.refine((answer) => {
+const quizAnswersSchema = z.object({
+	answers: z
+		.array(
+			quizAnswerSchema.refine((answer) => {
 				const { isSkipped, selectedOptions, userInput } = answer;
 
 				if (isSkipped) {
@@ -28,14 +18,28 @@ const QuizAnswersSchema = z.object({
 				}
 
 				const hasAnySelectedOptions =
-					selectedOptions.length > ZERO_ARRAY_LENGTH;
+					selectedOptions.length >= QuizAnswersRule.ANSWERS_LIST_MIN_LENGTH;
 				const hasUserInput = Boolean(userInput);
 
 				return hasAnySelectedOptions || hasUserInput;
 			}, QuizValidationMessage.NON_SKIPPED_QUESTION),
-	),
+		)
+		.refine(
+			(answers) =>
+				answers.filter((answer) => !answer.isSkipped).length >=
+				QuizAnswersRule.ANSWERS_LIST_MIN_LENGTH,
+			{
+				message: QuizValidationMessage.SKIPPED_QUESTIONS,
+			},
+		)
+		.refine(
+			(answers) => answers.length >= QuizAnswersRule.ANSWERS_LIST_MIN_LENGTH,
+			{
+				message: QuizValidationMessage.ANSWERS_REQUIRED,
+			},
+		),
 	category: z.nativeEnum(QuizCategory),
 	notes: z.string().trim().optional().default(""),
 });
 
-export { QuizAnswersSchema };
+export { quizAnswersSchema };